Adding QOL improvements, fixing xml layouts, etc
Added a new and better way to get the current user session data, made it so that if a group isn't assigned in create chore it doesn't create it, added proper image views for YourChores, added in chore completion (Though the point total doesn't show up in the sidebar, will have to fix later) plus a couple of other things I can't remember.

I have also added a testing branch to push unstable updates to the app now that we have a working prototype. We should use this in the future when we have improvements until we can all get it to work on our machines.